le-camera-server: Fix fd leak in continuous snapshot mode
Issue: In continuous snapshot mode, duplicate fds are getting
created in the client process which are not getting reused even after
closing. This is leading to too many open files error.

Fix: Similar to video track buffer handling, maintain a map of
original fd and duplicate fd to rotate the duplicate fds as well on
the client side and close the fds when the particular stream is closed.

Change-Id: Id2641611dd35711fdf3041ae8a7b2716192b59fd